stanli.sty: Use stanli's 3dplot defs only as fallback, load tikz-3dplot/3dplot packages if available.
stanli.sty provides a copy of some 3dplot definitions. However, the way it
is handled is incompatible with the use of {tikz-,}3dplot in the same
document, causing errors

! LaTeX Error: Command \tdplotsinandcos already defined.
               Or name \end... illegal, see p.192 of the manual.

This patch checks for tikz-3dplot/3dplot existence and requires them if
available. Since this part exits stanli.sty on success, stanli's own copy
of the definitions will only be used as fallback.

No problem if tikz-3dplot/3dplot package is already loaded, \RequirePackage
will not try to load it again.