stanli.sty:\support{6}: Generate a fake swirl with arc instead of using plot function.
Shape of torsional spring is achieved through a fake spiral made of arcs,
with slightly non-linear behavior.

It is scaled so its maximum radius matches \springTorsionRadius. This is a
new command to hold this maximum radius of torsional spring.

\springTorsionRadius default value is set to 6mm. This value is chosen to
match point to ground distance for bearings.

Ground location is set through \springTorsionRadius.

This commit also includes a clockwise variant of the torsional spring (6cw).